Move CaseJson into Caseload namespace
